+++ Changelog.md
@@ -1,5 +1,10 @@
 # YoutubeDLSharp Changelog
 
+### v.0.3 (2020-09)
+- **New:** Override options in OptionSet & YoutubeDL class.
+- **New:** More download progress information (transfer rate, ...)
+- **Fixed:** Converting option set to & from string.
+
 ### v.0.2 (2020-03)
 - **New:** Option to capture output in YoutubeDL class.
 - **New:** Added PythonPath to YoutubeDLProcess.